Initial stubs
Stubbing out the API based on the Mozilla Developer Network
documentation for the IndexedDB API.

Any method name with the word `delete` has been changed to destroy
since in AS3 delete is a keyword and will not compile, methods with
suffixes to the word delete have also been updated for consistency.
i.e. destroy() destroyDatabase()

Similarily, the method `continue` on the IDBCursor class has been
changed to `next`, since in AS3 continue is also a keyword and will nto
compile.